added parseSessionId to get guid out of session cookie
